Created a simple lottery that return balls with a random color and random number.
tasks which have done:

created Ball class with the following fields: color and number and override toString() method
created enum with different colors
implemented method getRandomColor() in class ColorSupplier. This method returns a random color, based on the enum values (use Random class for this)
Then created a Lottery class with getRandomBall() method, which returns the ball. This method creates a ball with a random color and a random number (the maximum possible number is 100) and return this ball from the method.

In main() method, created three balls using getRandomBall() method from Lottery class.